+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Change app name to "Codex Search". Applies in the application list in the Stripes menu-bar, within the application itself, and in the top-level route `codexsearch`. Fixes UISE-23.
 * New document, [the Codex contract](doc/codex-contract.md), which admittedly sounds like a Dan Brown book. Fixes UISE-28.
 * Dummy up the package file to provide an example of app metadata bundle. Fixes UISE-29, towards STCOR-117.
+* Display source-type icons instead of 'local', 'kb'. Fixes UISE-6.
 
 ## [1.1.0](https://github.com/folio-org/ui-search/tree/v1.1.0) (2017-12-05)
 [Full Changelog](https://github.com/folio-org/ui-search/compare/v1.0.0...v1.1.0)